+++ metrics/src/main/java/com/newrelic/telemetry/MetricBatchSender.java
@@ -14,6 +14,10 @@ import com.newrelic.telemetry.exceptions.RetryWithRequestedWaitException;
 import com.newrelic.telemetry.exceptions.RetryWithSplitException;
 import com.newrelic.telemetry.http.HttpPoster;
 import com.newrelic.telemetry.http.HttpResponse;
+import com.newrelic.telemetry.json.AttributesJson;
+import com.newrelic.telemetry.json.MetricBatchJson;
+import com.newrelic.telemetry.json.MetricJsonGenerator;
+import com.newrelic.telemetry.json.TelemetryBatchJson;
 import com.newrelic.telemetry.util.Utils;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
