+++ metrics/src/test/java/com/newrelic/telemetry/json/MetricBatchJsonTest.java
@@ -1,14 +1,16 @@
-package com.newrelic.telemetry;
+package com.newrelic.telemetry.json;
 
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import com.newrelic.telemetry.Attributes;
+import com.newrelic.telemetry.Gauge;
+import com.newrelic.telemetry.MetricBatch;
+import java.util.Collections;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
-
-import java.util.Collections;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 class MetricBatchJsonTest {
     private MetricJsonGenerator metricJsonGenerator;
